session: fix data race in the TestUpgradeVersion83 (#31889)
close pingcap/tidb#31831
